Remove trim step from connection.
Instead of calling trim as a separate step from connection, make it
transparent as part of the serialize call. Inside the packet the call to
the trim function in the payload no longer modifies the packet. Instead
if the payload needs to make the serialized packet be something other
than the current packet it creates a new one and returns it.
packet::serialize() will then use that packet for serialization.