Track min oob threshold in requests.
Add a field in request payloads to track the minimum oob threshold of
all the nodes the request has been routed through. This is used by the
node satisfying the request to determine if it should detach the content
from the start. It's preferable to detach right away if we know the
content will get detached anyways so that we know the peer that
satisfied the request, and is most likely to be holding the content
long-term, gets on the source list.