Hold the local node as a shared pointer.
Switch to having connections, protocols, and transports hold references
to the local node as a shared pointer instead of a plain reference. This
eliminates the need for a lot of ugly checking that has gone in to
prevent these classes from trying to access the node class after it has
been destroyed.